MyProfile/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
dj-database-url==0.5.0
Django==4.0.1
gunicorn==20.1.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2021.5
# whitenoise==5.3.0
django-tinymce==3.3.0
Pillow
djangorestframework
psycopg2-binary==2.9.1
